[level 2] Title: 더 맵게, Time: 1022.59 ms, Memory: 131 MB -BaekjoonHub

+++ 프로그래머스/2/42626. 더 맵게/더 맵게.java
@@ -3,19 +3,21 @@ class Solution {
     public int solution(int[] scoville, int K) {
         int answer = 0;
         // 1. min heap 만들기 
-        PriorityQueue<Long> minHeap = new PriorityQueue<>();
-        for (long s : scoville){
+        PriorityQueue<Integer> minHeap = new PriorityQueue<>();
+        for (int s : scoville){
             minHeap.offer(s);
         }
         
         // 2. min heap 맨 앞이 K보다 크거나 같을 때까지 새로운 음식 만들기(answer 세기)
         while(minHeap.size()>=2 && minHeap.peek() < K){
-            long min = minHeap.poll();
-            long nextMin = minHeap.poll();
-            long combi = min + (nextMin*2);
+            int min = minHeap.poll();
+            int nextMin = minHeap.poll();
+            int combi = min + (nextMin*2);
             minHeap.offer(combi);
             answer++;
         }
+        
+        // 3. 만들 수 없는 경우
         if (minHeap.peek() < K){
             return -1;
         }
